src/day4/requirements.txt
# 🚀 Day 4: Production Multi-Agent System Dependencies
# LangChain + CrewAI + Production Tools + Web Frameworks

# Core LLM and Agent Frameworks
langchain>=0.1.0
langchain-community>=0.0.1
langchain-experimental>=0.0.1
crewai>=0.1.0

# LLM Providers & Tools
transformers>=4.30.0
torch>=2.0.0
huggingface-hub>=0.16.0

# Vector Database & Memory
chromadb>=0.4.0
faiss-cpu>=1.7.0

# Tools & Integrations
duckduckgo-search>=3.8.0
python-dotenv>=1.0.0
requests>=2.31.0

# Data & File Handling
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0

# Configuration Management
pyyaml>=6.0
pydantic>=2.0.0

# Web Frameworks
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0

# API and Async Support
httpx>=0.25.0
aiofiles>=23.0.0

# Utilities
typing-extensions>=4.5.0
tqdm>=4.65.0

# Development & Visualization (for notebooks)
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Production Monitoring
psutil>=5.9.0

# Optional: For enhanced capabilities
# openai>=1.0.0  # If using OpenAI API
# anthropic>=0.3.0  # If using Claude API
